+++ source/Nuke.Common/Tools/AzureSignTool/AzureSignTool.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.AzureSignTool;
 /// <summary><p>Azure Sign Tool is similar to <c>signtool</c> in the Windows SDK, with the major difference being that it uses Azure Key Vault for performing the signing process. The usage is like <c>signtool</c>, except with a limited set of options for signing and options for authenticating to Azure Key Vault.</p><p>For more details, visit the <a href="https://github.com/vcsjones/AzureSignTool">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class AzureSignToolTasks : ToolTasks, IRequireNuGetPackage
 {