+++ source/Nuke.Common/Tools/CoverallsNet/CoverallsNet.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.CoverallsNet;
 /// <summary><p>Coveralls uploader for .Net Code coverage of your C# source code. Should work with any code files that get reported with the supported coverage tools, but the primary focus is CSharp.</p><p>For more details, visit the <a href="https://coverallsnet.readthedocs.io">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class CoverallsNetTasks : ToolTasks, IRequireNuGetPackage
 {