+++ source/Nuke.Common/Tools/Docker/Docker.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.Docker;
 /// <summary><p>Docker is an open platform for developing, shipping, and running applications. Docker enables you to separate your applications from your infrastructure so you can deliver software quickly. With Docker, you can manage your infrastructure in the same ways you manage your applications. By taking advantage of Docker’s methodologies for shipping, testing, and deploying code quickly, you can significantly reduce the delay between writing code and running it in production.</p><p>For more details, visit the <a href="https://www.docker.com/">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[PathToolRequirement(PathExecutable)]
 [PathTool(Executable = PathExecutable)]
 public partial class DockerTasks : ToolTasks, IRequirePathTool
 {