+++ source/Nuke.Common/Tools/DotCover/DotCover.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.DotCover;
 /// <summary><p>dotCover is a .NET unit testing and code coverage tool that works right in Visual Studio, helps you know to what extent your code is covered with unit tests, provides great ways to visualize code coverage, and is Continuous Integration ready. dotCover calculates and reports statement-level code coverage in applications targeting .NET Framework, Silverlight, and .NET Core.</p><p>For more details, visit the <a href="https://www.jetbrains.com/dotcover">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId)]
 public partial class DotCoverTasks : ToolTasks, IRequireNuGetPackage
 {