+++ source/Nuke.Common/Tools/InnoSetup/InnoSetup.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.InnoSetup;
 /// <summary><p>Inno Setup is a free installer for Windows programs by Jordan Russell and Martijn Laan. First introduced in 1997, Inno Setup today rivals and even surpasses many commercial installers in feature set and stability.</p><p>For more details, visit the <a href="http://www.jrsoftware.org/isinfo.php">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[PathToolRequirement(PathExecutable)]
 [PathTool(Executable = PathExecutable)]
 public partial class InnoSetupTasks : ToolTasks, IRequirePathTool
 {