+++ source/Nuke.Common/Tools/MakeNsis/MakeNSIS.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.MakeNSIS;
 /// <summary><p>NSIS creates installers that are capable of installing, uninstalling, setting system settings, extracting files, etc. Because it's based on script files you can fully control every part of your installer.</p><p>For more details, visit the <a href="https://nsis.sourceforge.io/Docs/Contents.html">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[PathToolRequirement(PathExecutable)]
 [PathTool(Executable = PathExecutable)]
 public partial class MakeNSISTasks : ToolTasks, IRequirePathTool
 {