+++ source/Nuke.Common/Tools/NUnit/NUnit.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.NUnit;
 /// <summary><p>NUnit is a unit-testing framework for all .Net languages. Initially ported from <a href="http://www.junit.org/">JUnit</a>, the current production release, version 3.0, has been completely rewritten with many new features and support for a wide range of .NET platforms.</p><p>For more details, visit the <a href="https://www.nunit.org/">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class NUnitTasks : ToolTasks, IRequireNuGetPackage
 {