+++ source/Nuke.Common/Tools/OpenCover/OpenCover.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.OpenCover;
 /// <summary><p>OpenCover is a code coverage tool for .NET 2 and above (Windows OSs only - no MONO), with support for 32 and 64 processes and covers both branch and sequence points.</p><p>For more details, visit the <a href="https://github.com/OpenCover/opencover">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class OpenCoverTasks : ToolTasks, IRequireNuGetPackage
 {