+++ source/Nuke.Common/Tools/Pwsh/Pwsh.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.Pwsh;
 /// <summary><p>PowerShell is a cross-platform task automation solution made up of a command-line shell, a scripting language, and a configuration management framework. PowerShell runs on Windows, Linux, and macOS.</p><p>For more details, visit the <a href="https://docs.microsoft.com/en-us/powershell/">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[PathToolRequirement(PathExecutable)]
 [PathTool(Executable = PathExecutable)]
 public partial class PwshTasks : ToolTasks, IRequirePathTool
 {