+++ source/Nuke.Common/Tools/SignClient/SignClient.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.SignClient;
 /// <summary><p>Code Signing client for Authenticode, NuGet, VSIX, and more</p><p>For more details, visit the <a href="https://discoverdot.net/projects/sign-service">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class SignClientTasks : ToolTasks, IRequireNuGetPackage
 {