+++ source/Nuke.Common/Tools/SonarScanner/SonarScanner.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.SonarScanner;
 /// <summary><p>The SonarScanner for MSBuild is the recommended way to launch a SonarQube or SonarCloud analysis for projects/solutions using MSBuild or dotnet command as build tool.</p><p>For more details, visit the <a href="https://www.sonarqube.org/">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class SonarScannerTasks : ToolTasks, IRequireNuGetPackage
 {