+++ source/Nuke.Common/Tools/TestCloud/TestCloud.Generated.cs
@@ -20,7 +20,6 @@ namespace Nuke.Common.Tools.TestCloud;
 /// <summary><p>Test Cloud is a cloud based service consisting of thousands of physical mobile devices. Users upload their apps and tests to Test Cloud, which will install the apps on the devices and run the tests. When the tests are complete, Test Cloud, the results made available to users through an easy to use and informative web-based front end.</p><p>For more details, visit the <a href="https://developer.xamarin.com/guides/testcloud/">official website</a>.</p></summary>
 [PublicAPI]
 [ExcludeFromCodeCoverage]
-[NuGetPackageRequirement(PackageId)]
 [NuGetTool(Id = PackageId, Executable = PackageExecutable)]
 public partial class TestCloudTasks : ToolTasks, IRequireNuGetPackage
 {