+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [vNext]
 
+## [9.0.2] / 2024-12-03
+- Fixed MSBuild tasks to use `net8.0` target framework
+- Fixed error handling in `:update` command
+- Fixed resolving tool path from options
+- Fixed nullable options for `ToolTasks.Run`
+- Fixed nullable underlying dictionary for delegate properties
+- Fixed skipping null and whitespace arguments
+- Fixed tool requirements
+- Fixed NPM tool path resolution
+- Fixed logging errors as standard in `GitTasks` and `DockerTasks`
+- Fixed argument format in `DotNetTasks`
+- Fixed nullable `Plugins` dictionary in `ReSharperTasks` 
+
 ## [9.0.1] / 2024-11-21
 - Fixed `Options` serialization to JSON
 - Fixed `Options` for default members in interfaces
@@ -1183,7 +1196,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added CLT tasks for Git
 - Fixed background color in console output
 
-[vNext]: https://github.com/nuke-build/nuke/compare/9.0.1...HEAD
+[vNext]: https://github.com/nuke-build/nuke/compare/9.0.2...HEAD
+[9.0.2]: https://github.com/nuke-build/nuke/compare/9.0.1...9.0.2
 [9.0.1]: https://github.com/nuke-build/nuke/compare/9.0.0...9.0.1
 [9.0.0]: https://github.com/nuke-build/nuke/compare/8.1.4...9.0.0
 [8.1.4]: https://github.com/nuke-build/nuke/compare/8.1.3...8.1.4